+++ R/GAGAs.R
@@ -13,9 +13,9 @@ NULL
 #> NULL
 
 
-#' Fit a GLM with GAGA algorithm
+#' Fit a GLM with the GAGA algorithm
 #'
-#' Fits linear, logistic and multinomial, poisson, and Cox regression models via Global Adaptive Generative Adjustment algorithm.
+#' Fits linear, logistic and multinomial, poisson, and Cox regression models via the Global Adaptive Generative Adjustment algorithm.
 #' @param X Input matrix, of dimension nobs*nvars; each row is an observation.
 #' If the intercept term needs to be considered in the estimation process, then the first column of \code{X} must be all 1s.
 #' @param y Response variable. Quantitative for \code{family="gaussian"}, or
@@ -40,6 +40,7 @@ NULL
 #' The running result of the algorithm is not sensitive to this value.
 #' @param fdiag It identifies whether to use diag Approximation to speed up the algorithm.
 #' @param frp Identifies if a method is preprocessed to reduce the number of parameters
+#' @param subItrNum Maximum number of steps for subprocess iterations.
 #'
 #' @return Regression coefficients
 #' @export GAGAs
@@ -50,9 +51,9 @@ NULL
 #' sample_size=300
 #' R1 = 3
 #' R2 = 2
-#' rate = 0.5 #Proportion of value zero in beta
-#' # Set true beta
-#' zeroNum = round(rate*p_size)
+#' ratio = 0.5 #The ratio of zeroes in coefficients
+#' # Set the true coefficients
+#' zeroNum = round(ratio*p_size)
 #' ind = sample(1:p_size,zeroNum)
 #' beta_true = runif(p_size,0,R2)
 #' beta_true[ind] = 0
@@ -80,9 +81,9 @@ NULL
 #' test_size=1000
 #' R1 = 1
 #' R2 = 3
-#' rate = 0.5 #Proportion of value zero in beta
-#' #Set true beta
-#' zeroNum = round(rate*p_size)
+#' ratio = 0.5 #The ratio of zeroes in coefficients
+#' #Set the true coefficients
+#' zeroNum = round(ratio*p_size)
 #' ind = sample(1:p_size,zeroNum)
 #' beta_true = runif(p_size,R2*0.2,R2)
 #' beta_true[ind] = 0
@@ -117,13 +118,13 @@ NULL
 #' classnames = c("C1","C2","C3","C4")
 #' sample_size = 500
 #' test_size = 1000
-#' rate = 0.5 #Proportion of value zero in beta
+#' ratio = 0.5 #The ratio of zeroes in coefficients
 #' Num = 10 # Total number of experiments
 #' R1 = 1
 #' R2 = 5
-#' #Set true beta
+#' #Set the true coefficients
 #' beta_true = matrix(rep(0,p_size*C),c(p_size,C))
-#' zeroNum = round(rate*p_size)
+#' zeroNum = round(ratio*p_size)
 #' for(jj in 1:C){
 #'   ind = sample(1:p_size,zeroNum)
 #'   tmp = runif(p_size,0,R2)
@@ -186,10 +187,10 @@ NULL
 #' sample_size=300
 #' R1 = 1/sqrt(p_size)
 #' R2 = 5
-#' rate = 0.5 #Proportion of value zero in beta
-#' # Set true beta
-#' zeroNum = round(rate*p_size)
-#' ind = sample(1:p_size,zeroNum)#'
+#' ratio = 0.5 #The ratio of zeroes in coefficients
+#' # Set the true coefficients
+#' zeroNum = round(ratio*p_size)
+#' ind = sample(1:p_size,zeroNum)
 #' beta_true = runif(p_size,0,R2)
 #' beta_true[ind] = 0
 #' X = R1*matrix(rnorm(sample_size * p_size), ncol = p_size)
@@ -208,11 +209,11 @@ NULL
 #' test_size = 1000
 #' R1 = 3
 #' R2 = 1
-#' rate = 0.5 #Proportion of value zero in beta
+#' ratio = 0.5 #The ratio of zeroes in coefficients
 #' censoringRate = 0.25 #Proportion of censoring data in observation data
-#' # Set true beta
-#' zeroNum = round(rate*p_size)
-#' ind = sample(1:p_size,zeroNum)#'
+#' # Set the true coefficients
+#' zeroNum = round(ratio*p_size)
+#' ind = sample(1:p_size,zeroNum)
 #' beta_true = runif(p_size,-R2,R2)
 #' beta_true[ind] = 0
 #' # Generate training samples
@@ -222,7 +223,7 @@ NULL
 #' t = ((-log(1-u)/(3*exp(z)))*100)^(0.1)
 #' cs = rep(0,sample_size)
 #' csNum = round(censoringRate*sample_size)
-#' ind = sample(1:sample_size,csNum)#'
+#' ind = sample(1:sample_size,csNum)
 #' cs[ind] = 1
 #' t[ind] = runif(csNum,0,0.8)*t[ind]
 #' y = cbind(t,1 - cs)
@@ -238,7 +239,7 @@ NULL
 #' t = ((-log(1-u)/(3*exp(z)))*100)^(0.1)
 #' cs = rep(0,test_size)
 #' csNum = round(censoringRate*test_size)
-#' ind = sample(1:test_size,csNum)#'
+#' ind = sample(1:test_size,csNum)
 #' cs[ind] = 1
 #' t[ind] = runif(csNum,0,0.8)*t[ind]
 #' y_t = cbind(t,1 - cs)
@@ -249,7 +250,7 @@ NULL
 #' cat("\n err:", norm(Eb-beta_true,type="2")/norm(beta_true,type="2"))
 #' cat("\n acc:", cal.w.acc(as.character(Eb!=0),as.character(beta_true!=0)))
 #' cat("\n Cindex:", cal.cindex(pred,y_t))
-GAGAs <- function(X,y,family=c("gaussian","binomial","poisson","multinomial","cox"),alpha=2,itrNum=500,thresh=1.e-3,QR_flag=FALSE,flag=TRUE,lamda_0=0.001,fdiag=TRUE,frp=TRUE) {
+GAGAs <- function(X,y,family=c("gaussian","binomial","poisson","multinomial","cox"),alpha=2,itrNum=100,thresh=1.e-3,QR_flag=FALSE,flag=TRUE,lamda_0=0.001,fdiag=TRUE,frp=TRUE,subItrNum = 20) {
 
   if(!is.character(family)){
     warning("Please check the input of family")
@@ -258,12 +259,12 @@ GAGAs <- function(X,y,family=c("gaussian","binomial","poisson","multinomial","co
 
   family=match.arg(family)
   beta=switch(family,
-             "gaussian"=LM_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,QR_flag=QR_flag,flag=flag,lamda_0=lamda_0,fix_sigma=FALSE,sigm2_0 = 1,fdiag=fdiag,frp=frp),
-             "poisson"=poisson_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,flag=flag,lamda_0=lamda_0,fdiag=fdiag),
-             "binomial"=logistic_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,flag=flag,lamda_0=lamda_0,fdiag=fdiag),
-             "multinomial"=multinomial_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,flag=flag,lamda_0=lamda_0,fdiag=fdiag),
-             "cox"=cox_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,flag=flag,lamda_0=lamda_0,fdiag=fdiag)
-            )
+              "gaussian"=LM_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,QR_flag=QR_flag,flag=flag,lamda_0=lamda_0,fix_sigma=FALSE,sigm2_0 = 1,fdiag=fdiag,frp=frp),
+              "poisson"=poisson_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,flag=flag,lamda_0=lamda_0,fdiag=fdiag,subItrNum),
+              "binomial"=logistic_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,flag=flag,lamda_0=lamda_0,fdiag=fdiag,subItrNum),
+              "multinomial"=multinomial_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,flag=flag,lamda_0=lamda_0,fdiag=fdiag,subItrNum),
+              "cox"=cox_GAGA(X,y,alpha=alpha,itrNum=itrNum,thresh=thresh,flag=flag,lamda_0=lamda_0,fdiag=fdiag,subItrNum)
+  )
 
   return(beta)
 }