+++ NEWS.md
@@ -1,3 +1,13 @@
+# GAGAs 0.6.1
+
+* GAGAs for linear models can return estimates of the standard deviation of the noise
+
+* Added cal.F1Score() function
+
+* Fix bug in case of no Coefficient selection
+
+* The maximum number of steps of subprocess iterations can be freely set
+
 # GAGAs 0.5
 
 * Used llt() instead of ldlt() to solve linear equations